smart-assistant/requirements-test.txt
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
fastapi[all]>=0.100.0
redis>=4.5.0
pymongo>=4.3.0
openai>=1.2.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.1.14
langchain_community>=0.0.10 